Remove leaking test helper
